Limit chest BE inventory sync to crystal only
Avoid BlockEntityData packet flood on generic chests (full inventory every change). Cuts network/client load near many chests; GUI still syncs via menu. Keep BE sync on crystal for top-item render; copy stacks in getTopStacks().
