+++ src/main/java/anner/ironchest/blocks/blockentities/CrystalChestEntity.java
@@ -1,10 +1,17 @@
 package anner.ironchest.blocks.blockentities;
 
 import anner.ironchest.blocks.ChestTypes;
-import net.minecraft.core.NonNullList;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.NonNullList;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.protocol.Packet;
+import net.minecraft.network.protocol.game.ClientGamePacketListener;
+import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.state.BlockState;
+import org.jspecify.annotations.Nullable;
 
 public class CrystalChestEntity extends GenericChestEntity {
 
@@ -19,11 +26,30 @@ public class CrystalChestEntity extends GenericChestEntity {
             if (stack.isEmpty()) {
                 continue;
             }
-            topStacks.set(itemCount++, stack);
+            topStacks.set(itemCount++, stack.copy());
             if (itemCount >= 12) {
                 break;
             }
         }
         return topStacks;
+    }
+
+    @Override
+    public void setChanged() {
+        super.setChanged();
+        if (this.level instanceof ServerLevel serverLevel) {
+            serverLevel.getChunkSource().blockChanged(this.worldPosition);
+        }
+    }
+
+    @Nullable
+    @Override
+    public Packet<ClientGamePacketListener> getUpdatePacket() {
+        return ClientboundBlockEntityDataPacket.create(this);
+    }
+
+    @Override
+    public CompoundTag getUpdateTag(HolderLookup.Provider registries) {
+        return this.saveWithoutMetadata(registries);
     }
 }
